StrmDrome/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0
python-dotenv>=1.0.0
pillow>=10.0.0
apscheduler>=3.10.0
bcrypt>=4.0.0
python-multipart>=0.0.9
mutagen>=1.47.0
